tn2dinf: add geometry.py file and GeometryInfinite2D.square constructor
