+++ docs/changelog.md
@@ -10,6 +10,7 @@ Release notes for `quimb`.
 - add [`CircuitPEPSSimpleUpdate`](#CircuitPEPSSimpleUpdate): a high level quantum circuit simulator that keeps the state as an arbitrary geometry PEPS and applies nearest-neighbor gates with 'simple update' style gauging. The geometry is given by a set of ``edges``, inferred from the ``gates`` or read from a ``psi0``; the accuracy is set by ``max_bond``; gauges can be periodically re-equilibrated with ``equilibrate``; local expectations are computed with the cluster approximation.
 - add [`CircuitPEPOSimpleUpdate`](#CircuitPEPOSimpleUpdate): the Heisenberg-picture companion to [`CircuitPEPSSimpleUpdate`](#CircuitPEPSSimpleUpdate). Gates are recorded lazily, then a local observable is built as a bond dimension 1 PEPO on the ``edges`` and evolved backwards through them with simple update gauging and compression (skipping gates outside its reverse lightcone), exposing the evolved operator via ``get_evolved_operator`` and its ``|00...0>`` expectation via ``local_expectation``.
 - add an infinite, translation-invariant 2D tensor-network family (`quimb.tensor.tn2dinf`) for infinite-PEPS ground states via simple update: [`GeometryInfinite2D`](#GeometryInfinite2D) defines a unit cell from a set of ``edges``, with sites labelled ``(cell, site_type)`` and bonds canonicalized into translation classes (``bond_type``); [`PEPSInfinite2D`](#PEPSInfinite2D) (and its base [`TensorNetworkInfinite2DFlat`](#TensorNetworkInfinite2DFlat)) stores a single tensor per site, keeping all translates in sync via shared tensors and indices; [`LocalHamInfinite2D`](#LocalHamInfinite2D) defines a Hamiltonian per ``bond_type`` (its geometry may extend beyond the state's, e.g. longer-range terms on a nearest-neighbor PEPS); and [`SimpleUpdateInfinite2D`](#SimpleUpdateInfinite2D) drives imaginary-time simple update by reusing [`SimpleUpdateGen`](#SimpleUpdateGen). Local expectations use a cluster (``max_distance``) or cheaper generalized-loop (``gloops``) approximation, and the dense, abelian-symmetric and fermionic (`symmray`) backends are all supported.
+- add [`GeometryInfinite2D.square`](#GeometryInfinite2D.square): build a square-lattice unit cell of any ``N`` by ``M`` size and neighbor range in one call, the range given as ``couplings`` (an int number of shells, ``1`` for nearest-neighbor, ``2`` to also include next-nearest, ..., or an explicit list of ``(dx, dy)`` site-step displacements) or a ``radius`` cutoff, both in basis-independent square-lattice distance, with sites placed at the 0-based lattice points.
 - [`TensorNetwork.gauge_all_simple`](#TensorNetwork.gauge_all_simple): add a ``fuse_multibonds`` option for updating gauges while preserving multi-index bonds, supported by explicit bond-index selection in [`tensor_compress_bond`](#tensor_compress_bond).
 - add [`tensor_gauge_simple_bond`](#tensor_gauge_simple_bond): the single-bond 'simple update' gauging step extracted from [`gauge_all_simple`](#TensorNetwork.gauge_all_simple), for gauging an individual bond between two tensors against a shared ``gauges`` dict.
 - [`tensor_network_ag_gate_simple`](#tensor_network_ag_gate_simple) (the ``gate_simple`` / ``gate_simple_`` methods): support long-range gates between sites that are not directly bonded, dispatching to the new [`tensor_network_ag_gate_simple_long_range`](#tensor_network_ag_gate_simple_long_range), which splits the gate into a matrix-product operator and applies it along a path of sites with simple-update gauging and recompression.
